seperate localStorage item grocery list to seperate module, update functions to use module
